abstract SessionStorage
