+++ lib/middleware/session_middleware.dart
@@ -1,0 +1,99 @@
+import 'dart:io';
+import 'dart:math';
+
+import 'package:shelf/shelf.dart';
+
+import '../session/session.dart';
+import 'cookies_middleware.dart';
+
+const _sessionKey = 'shelf_session.session_id';
+
+/// Returns the session middleware.
+Middleware sessionMiddleware() {
+  return (Handler innerHandler) {
+    return (Request request) async {
+      request = _addSessionIdToRequest(request);
+      final sessionId = getSessionId(request);
+      final requestedUri = request.requestedUri;
+      final isSecure = requestedUri.scheme == 'https';
+      final now = DateTime.now();
+      final lifetime = Session.lifetime;
+      final expires = now.add(lifetime);
+      final cookie = Cookie(
+        Session.name,
+        sessionId,
+      );
+      cookie.secure = isSecure;
+      cookie.path = '/';
+      cookie.maxAge = expires.difference(DateTime.now()).inSeconds;
+      cookie.expires = expires;
+      cookie.httpOnly = true;
+      request.addCookie(cookie);
+      final response = await innerHandler(request);
+      final session = Session.storage.getSession(sessionId);
+      if (session != null) {
+        session.expires = expires;
+      }
+
+      return response;
+    };
+  };
+}
+
+String getSessionId(Request request) {
+  final context = request.context;
+  if (!context.containsKey(_sessionKey)) {
+    throw StateError('The session id was not found in the request context');
+  }
+
+  return context[_sessionKey] as String;
+}
+
+Request _addSessionIdToRequest(Request request) {
+  final cookies = _parseCookieHeader(request);
+  var sessionId = cookies[Session.name];
+  sessionId ??= _generateSessionId();
+  request = request.change(context: {
+    _sessionKey: sessionId,
+  });
+  return request;
+}
+
+Map<String, String> _parseCookieHeader(Request request) {
+  final cookieHeader = request.headers[HttpHeaders.cookieHeader];
+  if (cookieHeader == null || cookieHeader.isEmpty) {
+    return <String, String>{};
+  }
+
+  try {
+    final result = <String, String>{};
+    for (final part in cookieHeader.split('; ')) {
+      final index = part.indexOf('=');
+      if (index != -1) {
+        result[part.substring(0, index)] = part.substring(index + 1);
+      }
+    }
+
+    return result;
+  } catch (s) {
+    return <String, String>{};
+  }
+}
+
+String _generateSessionId() {
+  while (true) {
+    final result = _getRandomString(32);
+    if (!Session.storage.sessionExist(result)) {
+      return result;
+    }
+  }
+}
+
+String _getRandomString(int length) {
+  const chars =
+      'AaBbCcDdEeFfGgHhIiJjKkLlMmNnOoPpQqRrSsTtUuVvWwXxYyZz1234567890';
+  final rnd = Random.secure();
+  return String.fromCharCodes(Iterable.generate(length, (_) {
+    return chars.codeUnitAt(rnd.nextInt(chars.length));
+  }));
+}
